# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Platform API",
    "description": "マルチユーザー対応AIエージェントプラットフォーム",
    "version": "0.1.0"
  },
  "paths": {
...
    "/webhooks/{webhook_path}": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "Handle Webhook",
        "description": "Webhookリクエストを受信してワークフロー実行.\n\n- HMAC署名が設定されている場合は検証を行う\n- ワークフローを非同期タスクとして実行",
        "operationId": "handle_webhook_webhooks__webhook_path__post",
        "parameters": [
          {
            "name": "webhook_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Webhook Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workflows/{workflow_id}/triggers/webhooks/{trigger_id}": {
      "delete": {
        "tags": [
          "triggers"
        ],
        "summary": "Delete Webhook Trigger",
        "description": "Webhookトリガーを削除.",
        "operationId": "delete_webhook_trigger_api_workflows__workflow_id__triggers_webhooks__trigger_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Workflow Id"
            }
          },
          {
            "name": "trigger_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trigger Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebhookErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "WebhookErrorResponse",
        "description": "Webhookエラーレスポンス."
      },
      "WebhookResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "workflow_id": {
            "type": "string",
            "title": "Workflow Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "task_id",
          "workflow_id"
        ],
        "title": "WebhookResponse",
        "description": "Webhook実行レスポンス."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}